@@ -3,6 +3,13 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
+	<parent>
+		<groupId>org.springframework.boot</groupId>
+		<artifactId>spring-boot-starter-parent</artifactId>
+		<version>3.3.3</version>
+		<relativePath/> <!-- lookup parent from repository -->
+	</parent>
+
 	<groupId>org.test</groupId>
 	<artifactId>commmon-eureka-server</artifactId>
 	<version>1</version>
@@ -11,32 +18,27 @@
 	<name>common-eureka-server</name>
 	<description>Eureka server suitable for use with most course labs</description>
 
-	<parent>
-		<groupId>org.springframework.boot</groupId>
-		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.0.1.RELEASE</version>
-		<relativePath/> <!-- lookup parent from repository -->
-	</parent>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<start-class>demo.Application</start-class>
-		<java.version>1.8</java.version>
+		<java.version>21</java.version>
+		<spring-cloud.version>2023.0.3</spring-cloud.version>		
 	</properties>
 
 	<!-- We can't have two parent POMs, so use DependencyManagement 
 	to show Maven where the Spring Cloud dependencies are. -->
-    <dependencyManagement>
-        <dependencies>
-	        <dependency>
-	            <groupId>org.springframework.cloud</groupId>
-	            <artifactId>spring-cloud-dependencies</artifactId>
-	            <version>Finchley.RELEASE</version>
-	            <type>pom</type>
-	            <scope>import</scope>
-	        </dependency>            
-        </dependencies>
-    </dependencyManagement>
+	<dependencyManagement>
+		<dependencies>
+		  <dependency>
+			<groupId>org.springframework.cloud</groupId>
+			<artifactId>spring-cloud-dependencies</artifactId>
+			<version>${spring-cloud.version}</version>
+			<type>pom</type>
+			<scope>import</scope>
+		  </dependency>
+		</dependencies>
+	</dependencyManagement>
 
 	<dependencies>
 	    <dependency>
