@@ -1,16 +1,14 @@
 package demo;
 
-import static org.junit.Assert.assertTrue;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import static org.assertj.core.api.Assertions.*;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
@@ -24,8 +22,8 @@ import org.springframework.web.context.WebApplicationContext;
  * 
  * @author ken krueger
  */
-@RunWith(SpringRunner.class)
-@SpringBootTest(classes = Application.class)
+
+@SpringBootTest
 @WebAppConfiguration
 @ActiveProfiles("native")	//	"native" means use local classpath location rather than GitHub.
 public class OutOfContainerTest {
@@ -33,7 +31,7 @@ public class OutOfContainerTest {
 	@Autowired WebApplicationContext spring;
 	MockMvc mockMvc;
 	
-	@Before
+	@BeforeEach
 	public void setup() {
 		mockMvc = MockMvcBuilders.webAppContextSetup(spring).build();
 	}
@@ -53,10 +51,9 @@ public class OutOfContainerTest {
 		
 		String returned = result.getResponse().getContentAsString();
 
+		
 		//	Check that the test values from the yml are present in the properties:
-		assertTrue(returned.contains("fromApplication:"));
-		assertTrue(returned.contains("applicationValue"));
-		assertTrue(returned.contains("fromTestConfig:"));
-		assertTrue(returned.contains("testConfigValue"));
+		assertThat(returned).contains("fromApplication:","applicationValue");
+		assertThat(returned).contains("fromTestConfig:","testConfigValue");
 	}
 }
