@@ -1,13 +1,12 @@
 package demo;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringRunner;
 
-@RunWith(SpringRunner.class)
-@SpringBootTest(classes = MyApplication.class)
+
+@SpringBootTest
 @ActiveProfiles("sentence")
 public class ApplicationTests {
 
