@@ -1,12 +1,11 @@
 package demo;
 
-import static org.junit.Assert.*;
+
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
@@ -24,8 +23,7 @@ import org.springframework.web.context.WebApplicationContext;
  * 
  * @author ken krueger
  */
-@RunWith(SpringRunner.class)
-@SpringBootTest(classes = Application.class)
+@SpringBootTest
 @WebAppConfiguration
 @ActiveProfiles("native")	//	"native" means use local classpath location rather than GitHub.
 public class OutOfContainerTest {
@@ -33,7 +31,7 @@ public class OutOfContainerTest {
 	@Autowired WebApplicationContext spring;
 	MockMvc mockMvc;
 	
-	@Before
+	@BeforeEach
 	public void setup() {
 		mockMvc = MockMvcBuilders.webAppContextSetup(spring).build();
 	}
