@@ -5,6 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.context.annotation.Bean;
+import org.springframework.web.client.RestClient;
 import org.springframework.web.client.RestTemplate;
 
 @SpringBootApplication
@@ -16,9 +17,11 @@ public class Application {
     }
 
     //	This "LoadBalanced" RestTemplate 
-    //	is automatically hooked into Ribbon:
-    @Bean @LoadBalanced
-    RestTemplate restTemplate() {
+    //	is automatically hooked into Spring Cloud's load balancer:
+    @Bean 
+    @LoadBalanced
+    public RestTemplate restTemplate() {
         return new RestTemplate();
     }    
+
 }
