@@ -1,7 +1,7 @@
 package demo;
 
 import java.net.URI;
-import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.client.ServiceInstance;
@@ -17,6 +17,7 @@ import org.springframework.web.client.RestTemplate;
 public class SentenceController {
 
 	@Autowired DiscoveryClient client;
+	RestTemplate template = new RestTemplate();
 	
 	
 	/**
@@ -60,14 +61,20 @@ public class SentenceController {
 	 * of speech is indicated by the given service / client ID:
 	 */
 	public String getWord(String service) {
-        List<ServiceInstance> list = client.getInstances(service);
-        if (list != null && list.size() > 0 ) {
-      	URI uri = list.get(0).getUri();
-	      	if (uri !=null ) {
-	      		return (new RestTemplate()).getForObject(uri,String.class);
-	      	}
-        }
-        return null;
+		Optional<URI> uri = client.getInstances(service)
+			.stream().findAny()
+			.map(ServiceInstance::getUri);
+
+		if (uri.isEmpty()) {
+			return "no " + service + " instances found.";
+		}
+
+		try {
+			return template.getForObject(uri.get(), String.class);
+		} catch (Exception e) {
+			System.out.println("Error retrieving " + service + " Error: " + e.getMessage());
+			return "(Error retrieving " + service + ")";
+		}
 	}
 
 }
