@@ -1,19 +1,21 @@
 package demo;
 
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
+import java.util.Random;
 
-@Controller
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
 public class WordController {
 
-	@Value("${words}") String words;
+    private final Random random = new Random();
+	
+	@Value("${words}") String[] words;
 	
 	@GetMapping("/")
-	public @ResponseBody String getWord() {
-		String[] wordArray = words.split(",");
-		int i = (int)Math.round(Math.random() * (wordArray.length - 1));
-		return wordArray[i];
+	public String getWord() {
+		return words[random.nextInt(words.length)];
 	}
+
 }
