@@ -2,20 +2,17 @@ package demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.zuul.EnableZuulProxy;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.filter.ShallowEtagHeaderFilter;
 
 @SpringBootApplication
-@EnableZuulProxy
+@EnableDiscoveryClient
 public class Application {
 
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
     }
     
-    @Bean
-    public ShallowEtagHeaderFilter shallowEtagHeaderFilter() {
-        return new ShallowEtagHeaderFilter();
-    }    
+
 }
